+++ tests/stdlib/tmemlinesBuf.nim
@@ -5,7 +5,7 @@ disabled: "appveyor"
 
 import memfiles
 var inp = memfiles.open("tests/stdlib/tmemlinesBuf.nim")
-var buffer: TaintedString = ""
+var buffer: string = ""
 var lineCount = 0
 for line in lines(inp, buffer):
   lineCount += 1