+++ changelog.md
@@ -285,6 +285,7 @@ proc mydiv(a, b): int {.raises: [].} =
 
 - Added a ``.noalias`` pragma. It is mapped to C's ``restrict`` keyword for the increased
   performance this keyword can enable.
+- Deprecated `TaintedString` and Taint mode.
 
 
 ## Compiler changes